LayoutTests/editing/editability/attr-invalid-string-expected.txt

When contentEditable attribute is invalid string, element.contentEditable returns "inherit".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".





PASS document.getElementById("div").getAttribute("contentEditable") is "abc"
PASS document.getElementById("div").contentEditable is "inherit"
PASS document.getElementById("div").isContentEditable is false
PASS window.getComputedStyle(div, "").getPropertyValue("-webkit-user-modify") is "read-only"

PASS successfullyParsed is true

TEST COMPLETE
